Fixed typecast.

+++ src/util/algo/upgma_mc.cpp
@@ -193,7 +193,7 @@ double load_edges(EdgeVec& all_edges, EdgeList &edges, vector<Node> &nodes, Queu
 	std::unordered_map<uint64_t, EdgePtr> edge_map;
 	edge_map.reserve(edges.size());
 	for (EdgePtr i = edges.begin(); i != edges.end(); ++i)
-		edge_map[(i->n1 << 32) | i->n2] = i;
+		edge_map[(uint64_t(i->n1) << 32) | i->n2] = i;
 
 	double evalue = lambda;
 	message_stream << "Reading edges..." << endl;
@@ -210,10 +210,10 @@ double load_edges(EdgeVec& all_edges, EdgeList &edges, vector<Node> &nodes, Queu
 		}
 		else {
 			if (i >= j) std::swap(i, j);
-			const auto e = edge_map.find((i << 32) | j);
+			const auto e = edge_map.find((uint64_t(i) << 32) | j);
 			if (e == edge_map.end()) {
 				const EdgePtr f = edges.emplace(edges.end(), i, j, 1, evalue);
-				edge_map[(i << 32) | j] = f;
+				edge_map[(uint64_t(i) << 32) | j] = f;
 			}
 			else {
 				++e->second->count;
